[12143] Fix some problems detected by static code analysis
Conflicts:
	src/game/BattleGround/BattleGroundMgr.cpp
	src/game/ObjectMgr.cpp
	src/game/PathFinder.cpp

+++ src/game/BattleGround/BattleGroundMgr.cpp
@@ -1196,7 +1196,7 @@ void BattleGroundMgr::BuildPvpLogDataPacket(WorldPacket *data, BattleGround *bg)
     if (type)                                                // arena
     {
         // it seems this must be according to BG_WINNER_A/H and _NOT_ TEAM_INDEX_A/H
-        for(int i = 1; i >= 0; --i)
+        for (int8 i = 1; i >= 0; --i)
         {
             uint32 pointsLost = bg->m_ArenaTeamRatingChanges[i] < 0 ? abs(bg->m_ArenaTeamRatingChanges[i]) : 0;
             uint32 pointsGained = bg->m_ArenaTeamRatingChanges[i] > 0 ? bg->m_ArenaTeamRatingChanges[i] : 0;
@@ -1206,7 +1206,8 @@ void BattleGroundMgr::BuildPvpLogDataPacket(WorldPacket *data, BattleGround *bg)
             *data << uint32(0);                                 // Matchmaking Value
             DEBUG_LOG("rating change: %d", bg->m_ArenaTeamRatingChanges[i]);
         }
-        for(int i = 1; i >= 0; --i)
+
+        for (int8 i = 1; i >= 0; --i)
         {
             uint32 at_id = bg->m_ArenaTeamIds[i];
             ArenaTeam * at = sObjectMgr.GetArenaTeamById(at_id);
